nutrition_ai_bot/requirements.txt
fastapi
uvicorn
pandas
numpy
faiss-cpu
sentence-transformers
requests
python-dotenv
